Adapt e2e compatibility tests to permissisonless ICS changes
